// shared/types.ts
export interface Member {
  slackID: string;
  name: string;
  kana: string;
  grade: string;
}